Extract OAuth flows from authentication.py
- Extract OAuthCallbackHandler, OAuthFlow, GoogleGeminiOAuthFlow, AnthropicOAuthFlow to oauth_flows.py
- Reduce authentication.py from 886 to 510 lines (376 line reduction)
- Maintain backward compatibility through re-exports and __all__ list
- All authentication tests passing
- Clean separation of OAuth flows from credential storage
