+++ src/llm_orc/authentication.py
@@ -1,18 +1,31 @@
 """Authentication system for LLM Orchestra supporting credential storage."""
 
 import os
-import secrets
-import threading
 import time
 import webbrowser
-from http.server import BaseHTTPRequestHandler, HTTPServer
 from typing import Any
-from urllib.parse import parse_qs, urlencode, urlparse
 
 import yaml
 from cryptography.fernet import Fernet
 
 from llm_orc.config import ConfigurationManager
+from llm_orc.core.auth.oauth_flows import (
+    AnthropicOAuthFlow,
+    GoogleGeminiOAuthFlow,
+    OAuthCallbackHandler,
+    OAuthFlow,
+    create_oauth_flow,
+)
+
+__all__ = [
+    "CredentialStorage",
+    "AuthenticationManager",
+    "OAuthCallbackHandler",
+    "OAuthFlow",
+    "GoogleGeminiOAuthFlow",
+    "AnthropicOAuthFlow",
+    "create_oauth_flow",
+]
 
 
 class CredentialStorage:
@@ -209,385 +222,6 @@ class CredentialStorage:
             self._save_credentials(credentials)
 
 
-class OAuthCallbackHandler(BaseHTTPRequestHandler):
-    """HTTP handler for OAuth callback."""
-
-    def do_GET(self) -> None:  # noqa: N802
-        """Handle GET request for OAuth callback."""
-        parsed_url = urlparse(self.path)
-        query_params = parse_qs(parsed_url.query)
-
-        # Store the authorization code
-        if "code" in query_params:
-            self.server.auth_code = query_params["code"][0]  # type: ignore
-            self.send_response(200)
-            self.send_header("Content-type", "text/html")
-            self.end_headers()
-            self.wfile.write(b"""
-            <html>
-            <body>
-            <h1>Authorization Successful!</h1>
-            <p>You can close this window and return to the CLI.</p>
-            </body>
-            </html>
-            """)
-        elif "error" in query_params:
-            self.server.auth_error = query_params["error"][0]  # type: ignore
-            self.send_response(400)
-            self.send_header("Content-type", "text/html")
-            self.end_headers()
-            self.wfile.write(
-                b"""
-            <html>
-            <body>
-            <h1>Authorization Failed</h1>
-            <p>Error: """
-                + query_params["error"][0].encode()
-                + b"""</p>
-            </body>
-            </html>
-            """
-            )
-        else:
-            self.send_response(400)
-            self.send_header("Content-type", "text/html")
-            self.end_headers()
-            self.wfile.write(b"Invalid callback")
-
-    def log_message(self, format_str: str, *args: Any) -> None:
-        """Suppress log messages."""
-        pass
-
-
-class OAuthFlow:
-    """Handles OAuth flow for LLM providers."""
-
-    def __init__(
-        self,
-        provider: str,
-        client_id: str,
-        client_secret: str,
-        redirect_uri: str = "http://localhost:8080/callback",
-    ):
-        self.provider = provider
-        self.client_id = client_id
-        self.client_secret = client_secret
-        self.redirect_uri = redirect_uri
-        self.state = secrets.token_urlsafe(32)
-
-    def get_authorization_url(self) -> str:
-        """Get the authorization URL for the provider."""
-        # This is a generic implementation - providers would override this
-        if self.provider == "google":
-            return (
-                f"https://accounts.google.com/o/oauth2/v2/auth?"
-                f"client_id={self.client_id}&"
-                f"redirect_uri={self.redirect_uri}&"
-                f"response_type=code&"
-                f"scope=https://www.googleapis.com/auth/userinfo.email&"
-                f"state={self.state}"
-            )
-        elif self.provider == "github":
-            return (
-                f"https://github.com/login/oauth/authorize?"
-                f"client_id={self.client_id}&"
-                f"redirect_uri={self.redirect_uri}&"
-                f"state={self.state}&"
-                f"scope=user:email"
-            )
-        else:
-            raise ValueError(f"OAuth not supported for provider: {self.provider}")
-
-    def start_callback_server(self) -> tuple[HTTPServer, int]:
-        """Start the callback server and return auth code."""
-        # Find an available port
-        port = 8080
-        while port < 8090:
-            try:
-                server = HTTPServer(("localhost", port), OAuthCallbackHandler)
-                server.auth_code = None  # type: ignore
-                server.auth_error = None  # type: ignore
-                break
-            except OSError:
-                port += 1
-        else:
-            raise RuntimeError("No available port for OAuth callback")
-
-        # Update redirect URI with actual port
-        self.redirect_uri = f"http://localhost:{port}/callback"
-
-        def run_server() -> None:
-            server.timeout = 1
-            while server.auth_code is None and server.auth_error is None:  # type: ignore
-                server.handle_request()
-
-        server_thread = threading.Thread(target=run_server)
-        server_thread.daemon = True
-        server_thread.start()
-
-        return server, port
-
-    def exchange_code_for_tokens(self, auth_code: str) -> dict[str, Any]:
-        """Exchange authorization code for tokens."""
-        # This would typically make an HTTP request to the provider's token endpoint
-        # For now, return a mock response
-        return {
-            "access_token": f"mock_access_token_{auth_code[:10]}",
-            "refresh_token": f"mock_refresh_token_{auth_code[:10]}",
-            "expires_in": 3600,
-            "token_type": "Bearer",
-        }
-
-
-class GoogleGeminiOAuthFlow(OAuthFlow):
-    """OAuth flow specific to Google Gemini API."""
-
-    def __init__(self, client_id: str, client_secret: str):
-        super().__init__("google", client_id, client_secret)
-
-    def get_authorization_url(self) -> str:
-        """Get the authorization URL for Google Gemini API."""
-        params = {
-            "client_id": self.client_id,
-            "redirect_uri": self.redirect_uri,
-            "response_type": "code",
-            "scope": "https://www.googleapis.com/auth/generative-language.retriever",
-            "state": self.state,
-        }
-        return f"https://accounts.google.com/o/oauth2/v2/auth?{urlencode(params)}"
-
-    def exchange_code_for_tokens(self, auth_code: str) -> dict[str, Any]:
-        """Exchange authorization code for tokens with Google."""
-        # For now, return a mock response that satisfies the test
-        return {
-            "access_token": f"google_access_token_{auth_code[:10]}",
-            "refresh_token": f"google_refresh_token_{auth_code[:10]}",
-            "expires_in": 3600,
-            "token_type": "Bearer",
-        }
-
-
-class AnthropicOAuthFlow(OAuthFlow):
-    """OAuth flow specific to Anthropic API with improved user guidance."""
-
-    def __init__(self, client_id: str, client_secret: str):
-        # Use Anthropic's own callback endpoint to avoid Cloudflare protection
-        super().__init__(
-            "anthropic",
-            client_id,
-            client_secret,
-            "https://console.anthropic.com/oauth/code/callback",
-        )
-
-        # Generate PKCE parameters for secure OAuth flow
-        import base64
-        import hashlib
-
-        self.code_verifier = (
-            base64.urlsafe_b64encode(secrets.token_bytes(32))
-            .decode("utf-8")
-            .rstrip("=")
-        )
-        self.code_challenge = (
-            base64.urlsafe_b64encode(
-                hashlib.sha256(self.code_verifier.encode("utf-8")).digest()
-            )
-            .decode("utf-8")
-            .rstrip("=")
-        )
-
-    @classmethod
-    def create_with_guidance(cls) -> "AnthropicOAuthFlow":
-        """Create an Anthropic OAuth flow with interactive client ID setup."""
-        import webbrowser
-
-        print("🔧 Anthropic OAuth Setup")
-        print("=" * 50)
-        print("To set up Anthropic OAuth authentication, you need to:")
-        print()
-        print("1. Visit the Anthropic Console: https://console.anthropic.com")
-        print("2. Navigate to your organization settings or developer tools")
-        print("3. Create an OAuth application/client")
-        print(
-            "4. Set the redirect URI to: https://console.anthropic.com/oauth/code/callback"
-        )
-        print("5. Copy the client ID and client secret")
-        print()
-
-        # Offer to open the console automatically
-        open_browser = (
-            input("Would you like to open the Anthropic Console now? (y/N): ")
-            .strip()
-            .lower()
-        )
-        if open_browser in ["y", "yes"]:
-            webbrowser.open("https://console.anthropic.com")
-            print("✅ Opened Anthropic Console in your browser")
-            print()
-
-        # Get client ID and secret from user
-        print("Enter your OAuth credentials:")
-        client_id = input("Client ID: ").strip()
-        if not client_id:
-            raise ValueError("Client ID is required")
-
-        client_secret = input("Client Secret: ").strip()
-        if not client_secret:
-            raise ValueError("Client Secret is required")
-
-        return cls(client_id, client_secret)
-
-    def get_authorization_url(self) -> str:
-        """Get the authorization URL for Anthropic API with validated parameters."""
-        params = {
-            "client_id": self.client_id,
-            "redirect_uri": self.redirect_uri,
-            "response_type": "code",
-            "state": self.state,
-            "scope": "org:create_api_key user:profile user:inference",
-            "code_challenge": self.code_challenge,
-            "code_challenge_method": "S256",
-        }
-        return f"https://claude.ai/oauth/authorize?{urlencode(params)}"
-
-    def start_manual_callback_flow(self) -> str:
-        """Start manual callback flow using Anthropic's own callback endpoint."""
-        print("🔧 Manual Authorization Code Extraction")
-        print("=" * 50)
-        print("Since we're using Anthropic's callback endpoint, you'll need to")
-        print("manually extract the authorization code from the redirect URL.")
-        print()
-        print("After completing OAuth authorization in your browser:")
-        print(
-            "1. You'll be redirected to: https://console.anthropic.com/oauth/code/callback"
-        )
-        print("2. Look at the URL in your browser's address bar")
-        print("3. Find the 'code' parameter in the URL")
-        print("4. Copy the authorization code value")
-        print()
-        print("Example URL:")
-        print(
-            "https://console.anthropic.com/oauth/code/callback?code=ABC123...&state=xyz"
-        )
-        print("                                                    ^^^^^^")
-        print("                                            (copy this code)")
-        print()
-
-        # Prompt user for the authorization code
-        while True:
-            auth_code = input(
-                "Enter the authorization code from the callback URL: "
-            ).strip()
-            if auth_code:
-                return auth_code
-            print("Please enter a valid authorization code.")
-
-    def exchange_code_for_tokens(self, auth_code: str) -> dict[str, Any]:
-        """Exchange authorization code for access and refresh tokens."""
-        import requests
-
-        data = {
-            "grant_type": "authorization_code",
-            "client_id": self.client_id,
-            "code": auth_code,
-            "code_verifier": self.code_verifier,
-            "redirect_uri": self.redirect_uri,
-        }
-
-        try:
-            print("🔄 Attempting token exchange request...")
-            print("   Endpoint: https://console.anthropic.com/v1/oauth/token")
-            print("   Method: POST")
-            print(f"   Data keys: {list(data.keys())}")
-
-            response = requests.post(
-                "https://console.anthropic.com/v1/oauth/token",
-                json=data,
-                headers={"Content-Type": "application/json"},
-                timeout=30,
-            )
-
-            print(f"   Response status: {response.status_code}")
-
-            if response.status_code == 200:
-                print("✅ Token exchange successful!")
-                return response.json()  # type: ignore[no-any-return]
-            else:
-                print(f"❌ Token exchange failed: {response.status_code}")
-                print("   This is expected due to Cloudflare protection")
-                return self._fallback_to_browser_instructions(auth_code)
-
-        except requests.exceptions.RequestException as e:
-            print(f"❌ Network error during token exchange: {e}")
-            print("   Falling back to browser-based token extraction")
-            return self._fallback_to_browser_instructions(auth_code)
-        except Exception as e:
-            print(f"❌ Unexpected error during token exchange: {e}")
-            return self._fallback_to_browser_instructions(auth_code)
-
-    def _fallback_to_browser_instructions(self, auth_code: str) -> dict[str, Any]:
-        """Provide instructions for manual token extraction when API exchange fails"""
-        print("\n" + "=" * 60)
-        print("🔧 MANUAL TOKEN EXTRACTION REQUIRED")
-        print("=" * 60)
-        print("The OAuth token endpoint is protected by Cloudflare.")
-        print("Please extract tokens manually using one of these methods:")
-        print()
-        print("METHOD 1: Browser Developer Tools")
-        print("1. Open https://console.anthropic.com in a new tab")
-        print("2. Open Developer Tools (F12 or Cmd+Option+I)")
-        print("3. Go to Application tab > Local Storage > console.anthropic.com")
-        print("4. Look for keys containing 'token', 'auth', or 'access'")
-        print("5. Copy any access tokens you find")
-        print()
-        print("METHOD 2: API Key Alternative")
-        print("1. Go to https://console.anthropic.com/settings/keys")
-        print("2. Create a new API key")
-        print("3. Use the API key instead of OAuth tokens")
-        print()
-        print(f"Your authorization code (for reference): {auth_code[:20]}...")
-        print("=" * 60)
-
-        # Return a structure indicating manual extraction is needed
-        return {
-            "requires_manual_extraction": True,
-            "auth_code": auth_code,
-            "instructions": "Manual token extraction required - see console output",
-            "alternative_url": "https://console.anthropic.com/settings/keys",
-        }
-
-    def validate_credentials(self) -> bool:
-        """Validate OAuth credentials by testing the authorization URL."""
-        try:
-            auth_url = self.get_authorization_url()
-            # Try to access the authorization URL to validate the client_id
-            import urllib.request
-
-            urllib.request.urlopen(auth_url, timeout=10)  # noqa: F841
-            # A 200 response indicates the endpoint is accessible
-            # A 403 might mean the endpoint exists but requires authentication
-            # Both are acceptable for validation purposes
-            return True
-        except urllib.error.HTTPError as e:
-            # 403 Forbidden might mean the endpoint exists but needs authentication
-            # This is still considered valid for OAuth setup
-            if e.code == 403:
-                return True
-            print(f"❌ OAuth validation failed: HTTP {e.code}")
-            return False
-        except Exception as e:
-            print(f"❌ OAuth validation failed: {e}")
-            return False
-
-
-def create_oauth_flow(provider: str, client_id: str, client_secret: str) -> OAuthFlow:
-    """Factory function to create the appropriate OAuth flow for a provider."""
-    if provider == "google":
-        return GoogleGeminiOAuthFlow(client_id, client_secret)
-    elif provider == "anthropic":
-        return AnthropicOAuthFlow(client_id, client_secret)
-    else:
-        raise ValueError(f"OAuth not supported for provider: {provider}")
 
 
 class AuthenticationManager:
