+++ src/llm_orc/core/auth/__init__.py
@@ -1,1 +1,17 @@
-"""Authentication and credential management."""
+"""OAuth authentication components."""
+
+from llm_orc.core.auth.oauth_flows import (
+    AnthropicOAuthFlow,
+    GoogleGeminiOAuthFlow,
+    OAuthCallbackHandler,
+    OAuthFlow,
+    create_oauth_flow,
+)
+
+__all__ = [
+    "OAuthCallbackHandler",
+    "OAuthFlow",
+    "GoogleGeminiOAuthFlow", 
+    "AnthropicOAuthFlow",
+    "create_oauth_flow",
+]
